Revamp threading mechanism
Instead of spawning a bunch of API requests concurrently and waiting for
them all to complete, split the entire page range into a bunch of
independent threads that each process a big chunk of values
autonomously. Since xapian doesn't like threading, we have to lock it.
Also lock I/O to prevent garbage output.

This is a massive speedup because it eliminates waiting for the slowest
path to complete.

The splitting algorithm is fairly naive and just repeatedly chops the
PostSet in half. This has very good splitting properties, i.e. in the
ideal case we end up with O(threads) size increase rather than O(posts).
The downside is that it totally destroys the chronological order on
posts. It remains to be seen whether this will be a major drawback.