Update database representation
Instead of storing (Successes, Failures), which is essentially paying
the same work twice, it's much more efficient to just record the
Successes and also store (as a separate intervalset) the total range of
values that have been scraped so far.

This way we're essentially just storing the places of all the
success/failure transitions, since the total scraped range will be more
or less trivial to store.